requirements.txt
Flask==3.1.0
openai==1.72.0
svgwrite==1.4.3
Werkzeug==3.1.3
dotenv~=0.9.9
python-dotenv~=1.1.0
svgpathtools~=1.6.1
protobuf~=5.29.4
pdf2image~=1.17.0
google.generativeai
spire~=0.4.2
CairoSVG~=2.8.2
pillow~=11.2.1
PyPDF2~=3.0.1
google~=3.0.0
h11~=0.16.0
defusedxml~=0.7.1
numpy~=2.2.6
typing_extensions~=4.13.2
pip~=23.2.1
pathlib~=1.0.1
configparser~=7.2.0
distro~=1.9.0
filelock~=3.18.0
rsa~=4.9.1
pyasn1~=0.6.1
grpcio~=1.71.0
requests~=2.32.3
sniffio~=1.3.1
MarkupSafe~=3.0.2
click~=8.2.1
Jinja2~=3.1.6
blinker~=1.9.0
itsdangerous~=2.2.0
httpcore~=1.0.9
idna~=3.10
httpx~=0.28.1
jiter~=0.10.0
charset-normalizer~=3.4.2
pytz~=2025.2
cffi~=1.17.1
proto-plus~=1.26.1
scipy~=1.15.3
pyasn1_modules~=0.4.2
cachetools~=5.5.2
certifi~=2025.4.26
urllib3~=2.4.0
packaging~=25.0
google-api-python-client~=2.169.0
grpcio-status~=1.71.0
pydantic~=2.11.4
pandas~=2.2.3
httplib2~=0.22.0
tqdm~=4.67.1
anyio~=4.9.0
pydantic_core~=2.33.2
pyparsing~=3.2.3
annotated-types~=0.7.0
simplejson~=3.20.1
uritemplate~=4.1.1
google-auth-httplib2~=0.2.0